apply clippy suggestions

+++ cspuz_solver_backend/src/puzzle/guidearrow.rs
@@ -10,8 +10,8 @@ pub fn solve(url: &str) -> Result<Board, &'static str> {
     let width = clues[0].len();
     let mut board = Board::new(BoardKind::Grid, height, width, is_unique(&ans));
     board.push(Item::cell(
-        ty as usize,
-        tx as usize,
+        ty,
+        tx,
         "black",
         ItemKind::Circle,
     ));
